+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-7183 Showing storage IDs when there are multiple storages with the same name
 * VFS-7202 Updated "bad data" backend error translation
 * VFS-6802 Updated commons
 * VFS-7042 Improved UX of storage import counters